+++ shuup/front/__init__.py
@@ -23,7 +23,8 @@ class ShuupFrontAppConfig(AppConfig):
         ],
         "admin_shop_form_part": [
             "shuup.front.admin_module.sorts_and_filters.form_parts.ConfigurationShopFormPart",
-            "shuup.front.admin_module.checkout.form_parts.CheckoutShopFormPart"
+            "shuup.front.admin_module.checkout.form_parts.CheckoutShopFormPart",
+            "shuup.front.admin_module.companies.form_parts.RegistrationSettingsFormPart"
         ],
         "notify_event": [
             "shuup.front.notify_events:OrderReceived",